lowfs: tar works
